feat(pinner): overhaul MSW test harness with stateful stores and orval mock generation
Replace hand-written global-state MSW handlers with a stateful store
architecture built on a generic MapStore<K,V> base class.

- Enable orval MSW mock generation (mock: { type: "msw" })
- Add @lumeweb/pinner/mocks package export via tsdown
- Introduce MapStore<K,V> with CRUD + reset, extended by PinStore,
  TusStore, OperationStore, WebsiteStore, IPNSStore
- Replace per-endpoint error handler exports with parameterized
  factories (createUnauthorizedHandler, createNotFoundHandler, etc.)
- Inline applyMockDelay/createMockUploadResult into upload-handlers.ts
- Delete msw-custom-handlers.ts (unused)
- Migrate all integration test files and int-test.ts fixtures
- Delete old handler files (msw-handlers.ts, msw-pin-service-adapter.ts,
  msw-upload-handlers.ts, msw-websites-ipns-handlers.ts,
  msw-error-handlers.ts, msw-custom-handlers.ts, msw-helpers.ts,
  pin/msw-handlers.ts, upload/msw-handlers.ts)
