+++ libs/pinner/src/pin/__tests__/int-test.ts
@@ -1,8 +1,12 @@
 import { createIntTest } from "@/__tests__/create-int-test";
-import { pinHandlers, resetPinServiceState } from "./msw-handlers";
+import { PinStore, createPinHandlers, resetPinServiceState } from "@/__tests__/msw";
+
+const pinStore = new PinStore();
+
+await pinStore.initializeDefaults();
 
 export const test = await createIntTest({
-  handlers: pinHandlers,
-  resetState: resetPinServiceState,
+  handlers: createPinHandlers(pinStore),
+  resetState: () => resetPinServiceState(pinStore),
   enableLogging: false,
 });
