--- libs/pinner/src/pin/__tests__/msw-handlers.ts
@@ -1,18 +1,0 @@
-// Pin module MSW handlers - re-exports from parent
-
-export {
-  createPinHandler,
-  listPinsHandler,
-  getPinHandler,
-  updatePinHandler,
-  deletePinHandler,
-  pinHandlers,
-  resetPinServiceState,
-} from "@/__tests__/msw-pin-service-adapter";
-export {
-  pinNotFoundHandler,
-  unauthorizedHandler,
-  rateLimitHandler,
-  serverErrorHandler,
-  createCustomPinHandlers,
-} from "@/__tests__/msw-handlers";
