+++ libs/pinner/src/upload/__tests__/int-test.ts
@@ -1,7 +1,11 @@
 import { createIntTest } from "@/__tests__/create-int-test";
-import { uploadHandlers } from "./msw-handlers";
+import { TusStore, OperationStore, createUploadHandlers, resetUploadState } from "@/__tests__/msw";
+
+const tusStore = new TusStore();
+const operationStore = new OperationStore();
 
 export const test = await createIntTest({
-  handlers: uploadHandlers,
+  handlers: createUploadHandlers(tusStore, operationStore),
+  resetState: () => resetUploadState(tusStore, operationStore),
   enableLogging: false,
 });
