--- libs/pinner/src/upload/__tests__/msw-handlers.ts
@@ -1,19 +1,0 @@
-// Upload module MSW handlers - re-exports from parent
-
-export {
-  tusOptionsHandler,
-  tusHeadHandler,
-  tusCreateHandler,
-  tusPatchHandler,
-  tusDeleteHandler,
-  tusUploadHandlers,
-  xhrUploadHandler,
-  xhrUploadHandlers,
-  accountUploadLimitHandler,
-  accountInfoHandler,
-  accountHandlers,
-  uploadHandlers,
-  uploadErrorHandler,
-  tusErrorHandler,
-  createCustomUploadHandlers,
-} from "../../__tests__/msw-handlers";
