+++ libs/pinner/tsdown.config.ts
@@ -3,7 +3,7 @@ import { createLibraryConfigWithExternals } from "@lumeweb/tsdown-config";
 
 export default defineConfig(
   createLibraryConfigWithExternals(
-    ["./src/index.ts", "!src/**/*.{spec,stories}.{ts,tsx}"],
+    ["./src/index.ts", "./src/api/generated/index.msw.ts", "!src/**/*.{spec,stories}.{ts,tsx}"],
     [/node_modules/, "@uppy/core", "@uppy/tus", "stream"]
   )
 );
